executor 파일 나누고 append redirection 처리
redirection infile, outfile 함수 나눴음(결국)
